+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 0.3.0 2018-11-23
 ### Added
-- Ability to open tabs from a list of urls
+- Ability to open tabs from a list of urls [PR #3](https://github.com/Meebuhs/tablist/pull/3)
 - Webpack build pipeline with sass support
+- Ability to copy and load from current window or all windows
+
+### Fixed
+- Fix some output copying issues [PR #2](https://github.com/Meebuhs/tablist/pull/2)
 
 ## 0.2.0 - 2018-11-13
 ### Added